requirements.txt
opencv-python
dt-apriltags
arena-py
scipy
numpy